requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
diskcache==5.6.1
docopt==0.6.2
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
html2markdown==0.1.7
idna==3.4
Jinja2==3.1.2
llama-cpp-python==0.1.77
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.25.2
openai-whisper==20230314
pipreqs==0.4.13
regex==2023.8.8
requests==2.31.0
soupsieve==2.4.1
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarg==0.1.9
